+++ samples/react-native/src/App.tsx
@@ -76,6 +76,12 @@ Sentry.init({
       didCallNativeInit,
     );
   },
+  _experiments: {
+    enableLogs: true,
+    beforeSendLog: (log) => {
+      return log;
+    },
+  },
   enableUserInteractionTracing: true,
   integrations(integrations) {
     integrations.push(